requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-dotenv==1.0.0
langchain==0.0.330
openai==1.2.4
pydantic==2.4.2
pytest==7.4.3
black==23.10.1
flake8==6.1.0
isort==5.12.0
supabase==1.0.3
email-validator==2.0.0.post2
httpx==0.24.1
python-multipart==0.0.6
jsonschema==4.18.0
ujson==5.8.0
ruff==0.0.277
